+++ ci/tests/test_base.py
@@ -1,4 +1,4 @@
-from flask_openapi.base import Swagger
+from flask_openapi.core import Swagger
 from flask_openapi.constants import DEFAULT_CONFIG, DEFAULT_ENDPOINT
 from flask_openapi.openapi.specs import get_apispecs
